requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.119.0
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
psycopg2-binary==2.9.11
pydantic==2.12.2
pydantic_core==2.41.4
python-dotenv==1.1.1
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.1
websockets==15.0.1
